Moved Fact to org.scalatest package.
