+++ src/main/resources/org/scalactic/ScalacticBundle.properties
@@ -38,11 +38,5 @@ hadLength={0} had length {1}
 hadSizeInsteadOfExpectedSize={0} had size {1} instead of expected size {2}
 hadSize={0} had size {1}
 commaBut={0}, but {1}
-leftParensCommaBut=({0}), but {1}
-rightParensCommaBut={0}, but ({1})
-bothParensCommaBut=({0}), but ({1})
 commaAnd={0}, and {1}
-leftParensCommaAnd=({0}), and {1}
-rightParensCommaAnd={0}, and ({1})
-bothParensCommaAnd=({0}), and ({1})
 variableWasValue={0} was {1}
