+++ src/main/resources/org/scalatest/ScalaTestBundle.properties
@@ -652,3 +652,10 @@ expectedNoErrorButGotTypeError=Expected no compiler error, but got the following
 expectedNoErrorButGotParseError=Expected no compiler error, but got the following parse error: "{0}", for code: {1}
 
 withFixtureNotAllowedInAsyncFixtures=withFixture is not allowed in AsyncFixtures
+
+leftParensCommaBut=({0}), but {1}
+rightParensCommaBut={0}, but ({1})
+bothParensCommaBut=({0}), but ({1})
+leftParensCommaAnd=({0}), and {1}
+rightParensCommaAnd={0}, and ({1})
+bothParensCommaAnd=({0}), and ({1})
